Visit delegation's qself under elided-infer lifetime rib too
