--- tests/ui/delegation/wrong-lifetime-rib-ice-156342.stderr
@@ -1,19 +1,0 @@
-error[E0423]: expected function, found unit variant `None`
-  --> $DIR/wrong-lifetime-rib-ice-156342.rs:8:11
-   |
-LL |     reuse None::<&()>;
-   |           ^^^^^^^^^^^ not a function
-
-error[E0116]: cannot define inherent `impl` for a type outside of the crate where the type is defined
-  --> $DIR/wrong-lifetime-rib-ice-156342.rs:6:1
-   |
-LL | impl Trait {
-   | ^^^^^^^^^^ impl for type defined outside of crate
-   |
-   = help: consider defining a trait and implementing it for the type or using a newtype wrapper like `struct MyType(ExternalType);` and implement it
-   = note: for more details about the orphan rules, see <https://doc.rust-lang.org/reference/items/implementations.html?highlight=orphan#orphan-rules>
-
-error: aborting due to 2 previous errors
-
-Some errors have detailed explanations: E0116, E0423.
-For more information about an error, try `rustc --explain E0116`.
